refactor: move stateful layer into luxcore
