+++ lib/LuxCore/src/LuxCore.jl
@@ -356,6 +356,9 @@ end
 preserves_state_type(l::Tuple) = all(preserves_state_type, l)
 preserves_state_type(l::NamedTuple) = all(preserves_state_type, values(l))
 
+include("stateful.jl")
+include("custom_layer_interface.jl")
+
 module Internal
 
     using Random: Xoshiro
